[update]BufferElementDataを使用してフェーズごとのエンティティ数を取得するように

+++ Assets/Scripts/Modules/EntityManagerModule.cs
@@ -1,6 +1,8 @@
 using Master.Configs;
 using System;
+using Unity.Collections;
 using Unity.Entities;
+using UnityEngine.Rendering;
 
 namespace Master.Entities
 {
@@ -11,15 +13,24 @@ namespace Master.Entities
             _world = world;
             _entityManager = world.EntityManager;
             _group = _world.CreateSystemManaged<ParticleSystemGroup>();
+            _phaseCountEntity = _entityManager.CreateEntity();
         }
 
         public void CreateSystems(int count, int kernelValue,
             PhaseConfigRepository phaseConfig)
         {
-            GlobalState globalState = new(count, kernelValue, phaseConfig.Phase1Configs, phaseConfig.Phase2Configs);
+            GlobalState globalState =
+                new(count, kernelValue, phaseConfig.Phase1Configs, phaseConfig.Phase2Configs);
             Entity globalStateEntity = _entityManager.CreateEntity(typeof(GlobalState));
             _entityManager.SetComponentData(globalStateEntity, globalState);
             _globalStateQuery = _entityManager.CreateEntityQuery(ComponentType.ReadOnly<GlobalState>());
+
+            // フェーズカウントの受け取りバッファを生成。
+            DynamicBuffer<PhaseCountEntity> buffer =
+                _entityManager.AddBuffer<PhaseCountEntity>(_phaseCountEntity);
+            buffer.ResizeUninitialized(kernelValue);
+            NativeArray<PhaseCountEntity> array = buffer.AsNativeArray();
+            for (int i = 0; i < kernelValue; i++) { array[i] = default; }
 
             SystemHandle initializeSystem = _world.CreateSystem<ParticleInitializeSystem>();
             SystemHandle phase1System = _world.CreateSystem<Phase1UpdateSystem>();
@@ -39,11 +50,25 @@ namespace Master.Entities
             _group.Update();
         }
 
-        public GlobalState GetGlobalState() => _globalStateQuery.GetSingleton<GlobalState>();
+        public void GetPhaseCount(Span<int> output)
+        {
+            if (!_entityManager.HasBuffer<PhaseCountEntity>(_phaseCountEntity)) { return; }
+
+            DynamicBuffer<PhaseCountEntity> buffer =
+                _entityManager.GetBuffer<PhaseCountEntity>(_phaseCountEntity);
+            NativeArray<PhaseCountEntity> array = buffer.AsNativeArray();
+            for (int i = 0; i < output.Length; i++)
+            {
+                output[i] = array[i].Count;
+            }
+        }
 
         public void Dispose()
         {
+            if (!_world.IsCreated) { return; }
+
             _entityManager.DestroyEntity(_globalStateQuery);
+            _entityManager.DestroyEntity(_phaseCountEntity);
 
             if (_group.Enabled)
             {
@@ -58,6 +83,7 @@ namespace Master.Entities
         private readonly World _world;
         private readonly EntityManager _entityManager;
         private readonly ParticleSystemGroup _group;
+        private readonly Entity _phaseCountEntity;
 
         private EntityQuery _globalStateQuery;
     }
